+++ lib/src/services/log_service/log_backend.dart
@@ -1,3 +1,5 @@
+import 'package:cl_datahub/cl_datahub.dart';
+
 import 'log_message.dart';
 
 /// Backend interface for [LogService].
@@ -7,4 +9,5 @@ import 'log_message.dart';
 abstract class LogBackend {
   Future<void> initialize() => Future.value();
   void publish(LogMessage message);
+  void setLogLevel(int level);
 }